+++ Training Project/App/SceneDelegate.swift
@@ -1,0 +1,38 @@
+//
+//  SceneDelegate.swift
+//  Training Project
+//
+//  Created by Daniil Litvinov on 18.06.2024.
+//
+
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+    
+    var window: UIWindow?
+    
+    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+        guard let windowScene = (scene as? UIWindowScene) else { return }
+        self.window = UIWindow(windowScene: windowScene)
+//        let navigationController = UINavigationController(rootViewController: MailViewController())
+        window?.rootViewController = MailViewController()
+        window?.makeKeyAndVisible()
+    }
+    
+    func sceneDidDisconnect(_ scene: UIScene) {
+    }
+    
+    func sceneDidBecomeActive(_ scene: UIScene) {
+    }
+    
+    func sceneWillResignActive(_ scene: UIScene) {
+    }
+    
+    func sceneWillEnterForeground(_ scene: UIScene) {
+    }
+    
+    func sceneDidEnterBackground(_ scene: UIScene) {
+        (UIApplication.shared.delegate as? AppDelegate)?.saveContext()
+    }
+}
+
