+++ Training Project/Managers/Training_Project.xcdatamodeld/.xccurrentversion
