+++ Training Project/Managers/Training_Project.xcdatamodeld/Training_Project.xcdatamodel/contents
