+++ Training Project/Moduls/MailScreen/MailViewController.swift
@@ -1,0 +1,26 @@
+//
+//  MailViewController.swift
+//  Training Project
+//
+//  Created by Daniil Litvinov on 19.06.2024.
+//
+
+import UIKit
+
+final class MailViewController: UIViewController {
+    //MARK: viewDidLoad
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        setupHerarchy()
+        setupView()
+    }
+    //MARK: - Setups
+    
+    private func setupView() {
+        view.backgroundColor = .red
+    }
+    
+    private func setupHerarchy() {
+    }
+    
+}
