+++ Training Project/Resourse/Assets.xcassets/AccentColor.colorset/Contents.json
