+++ Training Project/Resourse/Assets.xcassets/AppIcon.appiconset/Contents.json
