+++ Training Project/Resourse/Assets.xcassets/Contents.json
