+++ Training Project/Resourse/Info.plist
