--- Training Project/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  Training Project
-//
-//  Created by Daniil Litvinov on 18.06.2024.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
